refactor: change enum to an extern
